fix call api to use RequestOptions instead of map; move some api call functions from UpcloudStatusUtility back to UpcloudApiService
